Matmul via f16
