+++ mistralrs-core/src/models/phi2.rs
@@ -9,12 +9,11 @@ use candle_core::{quantized::QMatMul, DType, Device, Module, Result, Tensor};
 use candle_nn::{
     embedding, layer_norm, linear, Activation, Embedding, LayerNorm, RotaryEmbedding, VarBuilder,
 };
-use mistralrs_lora::layer::QLinear;
 use serde::Deserialize;
 
 use crate::{
     device_map::DeviceMapper,
-    layers::{flash_attn, repeat_kv, CausalMasker},
+    layers::{flash_attn, repeat_kv, CausalMasker, MatMul, QLinear},
     pipeline::{extract_logits, Cache, NormalModel},
     DeviceMapMetadata,
 };
@@ -118,7 +117,7 @@ impl Attention {
         } else {
             (None, None)
         };
-        let softmax_scale = 1f64 / (head_dim as f64).sqrt();
+        let softmax_scale = (head_dim as f64).sqrt();
         Ok(Self {
             q_proj: QLinear::from_linear(q_proj),
             k_proj: QLinear::from_linear(k_proj),
@@ -206,16 +205,13 @@ impl Attention {
             let v = v.transpose(1, 2)?;
             flash_attn(&q, &k, &v, self.softmax_scale as f32, seq_len > 1)?.transpose(1, 2)?
         } else {
-            let attn_weights = (q
-                .to_dtype(DType::F32)?
-                .contiguous()?
-                .matmul(&k.to_dtype(DType::F32)?.t()?)?
-                * self.softmax_scale)?;
+            let attn_weights =
+                MatMul.matmul_affine(&q.contiguous()?, &k.t()?, self.softmax_scale)?;
             let attn_weights =
                 CausalMasker.apply_mask(&mask.cloned(), attn_weights, &self.neg_inf)?;
             let attn_weights =
                 candle_nn::ops::softmax_last_dim(&attn_weights)?.to_dtype(v.dtype())?;
-            attn_weights.matmul(&v)?
+            MatMul.matmul(&attn_weights, &v)?
         };
         if self.q_proj.is_quant() {
             attn_output = attn_output.to_dtype(DType::F32)?;
