+++ mistralrs-core/src/xlora_models/llama.rs
@@ -1,17 +1,18 @@
 #![allow(clippy::cast_possible_truncation, clippy::cast_precision_loss)]
 
+use crate::{
+    layers::MatMul,
+    lora::{linear_no_bias as linear, LinearLayerLike, LoraConfig, Ordering},
+};
 use candle_core::{quantized::QMatMul, DType, Device, Result, Tensor, D};
 use candle_nn::{embedding, Embedding, Module, RotaryEmbedding, VarBuilder};
-use mistralrs_lora::{
-    layer::QLinear, linear_no_bias as linear, LinearLayerLike, LoraConfig, Ordering,
-};
 use std::{collections::HashMap, sync::Arc};
 use tqdm::Iter;
 use tracing::info;
 
 use crate::{
     device_map::DeviceMapper,
-    layers::{flash_attn, repeat_kv, CausalMasker, RmsNorm},
+    layers::{flash_attn, repeat_kv, CausalMasker, QLinear, RmsNorm},
     models::llama::Config,
     pipeline::{self, extract_logits, LayerCaches, NormalModel},
     DeviceMapMetadata,
@@ -113,15 +114,11 @@ impl CausalSelfAttention {
             let softmax_scale = 1f32 / (self.head_dim as f32).sqrt();
             flash_attn(&q, &k, &v, softmax_scale, seq_len > 1)?.transpose(1, 2)?
         } else {
-            let in_dtype = q.dtype();
-            let q = q.to_dtype(DType::F32)?;
-            let k = k.to_dtype(DType::F32)?;
-            let v = v.to_dtype(DType::F32)?;
-            let att = (q.matmul(&k.t()?)? / (self.head_dim as f64).sqrt())?;
+            let att = MatMul.matmul_affine(&q, &k.t()?, (self.head_dim as f64).sqrt())?;
             let att = CausalMasker.apply_mask(mask, att, &self.neg_inf)?;
             let att = candle_nn::ops::softmax(&att, D::Minus1)?;
             // Convert to contiguous as matmul doesn't support strided vs for now.
-            att.matmul(&v.contiguous()?)?.to_dtype(in_dtype)?
+            MatMul.matmul(&att, &v.contiguous()?)?
         };
         let mut y = y.transpose(1, 2)?.reshape(&[b_sz, seq_len, hidden_size])?;
         if self.q_proj.is_quant() {
