--- mistralrs-lora/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "mistralrs-lora"
-readme = "README.md"
-authors = ["Eric Buehler"]
-version.workspace = true
-edition.workspace = true
-description.workspace = true
-repository.workspace = true
-keywords.workspace = true
-categories.workspace = true
-license.workspace = true
-homepage.workspace = true
-
-[dependencies]
-candle-core.workspace = true
-candle-nn.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-either.workspace = true
-accelerate-src = { workspace = true, optional = true }
-intel-mkl-src = { workspace = true, optional = true }
-
-[features]
-cuda = ["candle-core/cuda", "candle-nn/cuda"]
-cudnn = ["candle-core/cudnn"]
-metal = ["candle-core/metal", "candle-nn/metal"]
-accelerate = ["candle-core/accelerate", "candle-nn/accelerate"]
-mkl = ["candle-core/mkl", "candle-nn/mkl"]
