--- mistralrs-lora/src/layer.rs
@@ -1,99 +1,0 @@
-use std::sync::Arc;
-
-use candle_core::{
-    quantized::{gguf_file, QMatMul, QTensor},
-    DType, Device, Result, Tensor,
-};
-use candle_nn::{Linear, Module};
-
-#[derive(Debug, Clone)]
-pub struct QLinear {
-    inner: QMatMul,
-    bias: Option<Tensor>,
-    dtype: DType,
-}
-
-impl QLinear {
-    pub fn new<R: std::io::Read + std::io::Seek>(
-        ct: &gguf_file::Content,
-        r: &mut R,
-        name: &str,
-        device: &Device,
-    ) -> Result<Self> {
-        let w = ct.tensor(r, &format!("{name}.weight"), device)?;
-        let b = ct.tensor(r, &format!("{name}.bias"), device)?;
-        let inner = QMatMul::from_qtensor(w)?;
-        let bias = b.dequantize(device)?;
-        Ok(Self {
-            inner,
-            bias: Some(bias),
-            dtype: DType::F32,
-        })
-    }
-
-    pub fn from_linear(linear: Linear) -> Self {
-        Self {
-            inner: QMatMul::Tensor(linear.weight().clone()),
-            bias: linear.bias().cloned(),
-            dtype: if linear.weight().device().is_cuda() {
-                DType::BF16
-            } else {
-                DType::F32
-            },
-        }
-    }
-
-    pub fn from_parts(w: Tensor, b: Option<Tensor>) -> Self {
-        let dtype = if w.device().is_cuda() {
-            DType::BF16
-        } else {
-            DType::F32
-        };
-        Self {
-            inner: QMatMul::Tensor(w),
-            bias: b,
-            dtype,
-        }
-    }
-
-    pub fn from_qparts(w: QTensor, b: Option<Tensor>) -> Self {
-        if let Some(ref b) = b {
-            assert_eq!(b.dtype(), DType::F32);
-        }
-        Self {
-            inner: QMatMul::QTensor(Arc::new(w)),
-            bias: b,
-            dtype: DType::F32,
-        }
-    }
-
-    pub fn inner(&mut self) -> &mut QMatMul {
-        &mut self.inner
-    }
-
-    pub fn is_quant(&self) -> bool {
-        matches!(self.inner, QMatMul::QTensor(_))
-    }
-
-    pub fn bias(&self) -> Option<&Tensor> {
-        self.bias.as_ref()
-    }
-}
-
-impl Module for QLinear {
-    fn forward(&self, xs: &Tensor) -> Result<Tensor> {
-        let xs = if self.is_quant() {
-            xs.to_dtype(DType::F32)?
-        } else {
-            xs.clone()
-        };
-        if let Some(bias) = &self.bias {
-            self.inner
-                .forward(&xs)?
-                .broadcast_add(bias)?
-                .to_dtype(self.dtype)
-        } else {
-            self.inner.forward(&xs)?.to_dtype(self.dtype)
-        }
-    }
-}
